add internals directory and organize
